Dice Rolling Simulator
Like the title suggests, this project involves writing a program that simulates rolling die. When the program runs, it will randomly choose a number between 1 and 6. The program will print what that number what the number is. It should then ask you if you would like to roll the die again.